refactor(attachment): move the offline blob pipeline out of dexie/ into offline/
Step 4 of .todos/ATTACHMENTS_REFACTOR.md.

`dexie/` was a misnomer in both directions: it named a folder after the IndexedDB wrapper
while holding the module's most load-bearing domain logic (storage-service, download-queue,
and the schema types query/app-db itself depends on), and the files inside barely touched
Dexie — the real liveQuery usage lived outside it in download-service.ts.

The pipeline is also one subsystem, so the two background services join their collaborators
instead of sitting at module root: download-queue.ts had exactly one production importer
(download-service.ts) a folder away, both services start/stop together in query/provider.tsx,
and tests/ already covers the five files as a unit.

- attachment/dexie/ -> attachment/offline/, with download-service.ts and upload-service.ts
  moved in from the module root
- dialog/handler.tsx -> dialog/attachment-dialog-handler.tsx and dialog/helpers.tsx ->
  dialog/open-attachment-dialog.tsx, matching the sibling convention (user-sheet-handler.tsx)
- table/cells.tsx -> table/attachment-cells.tsx; table/helpers.tsx ->
  table/use-attachments-upload-dialog.tsx, with the app-generic formatBytes moved to
  ~/utils/format-bytes (two other call sites hand-roll the same MB math and can now reuse it)

Adds cella/migrations/2026-07-attachment-offline-layout/ so forks can run the same sweep:
an allow-list codemod for the import paths and upload-vocabulary renames, which also reports
references to removed exports (they have no mechanical fix). Verified idempotent against this
already-migrated tree.
