+++ frontend/src/modules/attachment/offline/download-service.ts
@@ -2,10 +2,10 @@ import { onlineManager } from '@tanstack/react-query';
 import { liveQuery, type Subscription } from 'dexie';
 import type { Attachment } from 'sdk';
 import { appConfig } from 'shared';
-import { attachmentsDb } from '~/modules/attachment/dexie/attachments-db';
-import { downloadQueue, SKIP_REASON_NO_ORIGINAL_KEY } from '~/modules/attachment/dexie/download-queue';
-import { attachmentStorage } from '~/modules/attachment/dexie/storage-service';
 import { type CloudFileVariant, getCloudUrl, getVariantKey } from '~/modules/attachment/file-url';
+import { attachmentsDb } from '~/modules/attachment/offline/attachments-db';
+import { downloadQueue, SKIP_REASON_NO_ORIGINAL_KEY } from '~/modules/attachment/offline/download-queue';
+import { attachmentStorage } from '~/modules/attachment/offline/storage-service';
 import { attachmentQueryKeys, findAttachmentInCache } from '~/modules/attachment/query';
 import { isPersisted } from '~/modules/attachment/types';
 import { getAppDb } from '~/query/app-db';
