+++ frontend/src/modules/attachment/table/attachments-bar.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next';
 import type { Attachment } from 'sdk';
 import { DeleteAttachments } from '~/modules/attachment/delete-attachments';
 import type { AttachmentsTableProps } from '~/modules/attachment/table/attachments-table';
-import { useAttachmentsUploadDialog } from '~/modules/attachment/table/helpers';
+import { useAttachmentsUploadDialog } from '~/modules/attachment/table/use-attachments-upload-dialog';
 import type { AttachmentsRouteSearchParams } from '~/modules/attachment/types';
 import { AlertBanner } from '~/modules/common/alerter/alert-banner';
 import { ColumnsView } from '~/modules/common/data-table/columns-view';
