+++ frontend/src/modules/attachment/table/attachments-columns.tsx
@@ -3,8 +3,7 @@ import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { Attachment } from 'sdk';
 import { seenWindowMs } from 'shared';
-import { DownloadCell, EllipsisCell, ThumbnailCell } from '~/modules/attachment/table/cells';
-import { formatBytes } from '~/modules/attachment/table/helpers';
+import { DownloadCell, EllipsisCell, ThumbnailCell } from '~/modules/attachment/table/attachment-cells';
 import { EditCellInput } from '~/modules/common/data-grid/cell-renderers';
 import { CheckboxColumn } from '~/modules/common/data-table/checkbox-column';
 import type { ColumnOrColumnGroup } from '~/modules/common/data-table/types';
@@ -13,6 +12,7 @@ import { SeenMark } from '~/modules/seen/seen-mark';
 import { UserCell } from '~/modules/user/user-cell';
 import { cn } from '~/utils/cn';
 import { dateShort } from '~/utils/date-short';
+import { formatBytes } from '~/utils/format-bytes';
 
 /** Views are only counted inside the seen ledger's retention window; older rows read 0. */
 const isOutsideSeenWindow = (createdAt: string | null | undefined) => {
