+++ frontend/src/modules/attachment/table/use-attachments-upload-dialog.tsx
@@ -65,17 +65,3 @@ export const useAttachmentsUploadDialog = (tenantId: string, organizationId: str
 
   return { open };
 };
-
-export const formatBytes = (bytes: string): string => {
-  const parsedBytes = Number(bytes);
-
-  if (parsedBytes <= 0 || Number.isNaN(parsedBytes)) return '0 B';
-
-  const sizes = ['B', 'KB', 'MB', 'GB', 'TB'];
-  const index = Math.floor(Math.log(parsedBytes) / Math.log(1024));
-
-  // Show 2 decimal places for MB or higher, else round to whole number
-  const formattedSize = (parsedBytes / 1024 ** index).toFixed(index > 1 ? 2 : 0);
-
-  return `${formattedSize} ${sizes[index]}`;
-};
