+++ frontend/src/modules/common/blocknote/full-html.tsx
@@ -5,9 +5,9 @@ import '~/modules/common/blocknote/custom-elements/checklist/checklist-styles.cs
 import DOMPurify from 'dompurify';
 import { type MouseEventHandler, useEffect, useRef, useState } from 'react';
 import type { CarouselItemData } from '~/modules/attachment/attachments-carousel';
-import { attachmentStorage } from '~/modules/attachment/dexie/storage-service';
 import { openAttachmentDialog } from '~/modules/attachment/dialog/open-attachment-dialog';
 import { getPrivateFileUrlById, getPublicFileUrl } from '~/modules/attachment/file-url';
+import { attachmentStorage } from '~/modules/attachment/offline/storage-service';
 import { findAttachmentInCache } from '~/modules/attachment/query';
 import {
   findClickedMedia,
