+++ frontend/src/modules/common/uploader/helpers/uppy-helpers.ts
@@ -5,8 +5,8 @@ import Transloadit from '@uppy/transloadit';
 import { getUploadToken, type UploadToken } from 'sdk';
 import { appConfig } from 'shared';
 import { nanoid } from 'shared/utils/nanoid';
-import { makeBlobKey, type UploadContext } from '~/modules/attachment/dexie/attachments-db';
-import { attachmentStorage } from '~/modules/attachment/dexie/storage-service';
+import { makeBlobKey, type UploadContext } from '~/modules/attachment/offline/attachments-db';
+import { attachmentStorage } from '~/modules/attachment/offline/storage-service';
 import { prepareFilesForOffline } from '~/modules/common/uploader/helpers/prepare-for-offline';
 import type { CustomUppy, CustomUppyFile, CustomUppyOpt } from '~/modules/common/uploader/types';
 import type { UploadTokenQuery } from '~/modules/me/types';
